requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
httpx==0.24.0
python-dotenv==1.1.0